+++ packages/docs/src/locales/ru/ru.json
@@ -1006,8 +1006,8 @@
         "errorMessages": "Отображает сообщения или сообщение об ошибке, если используются данные типа `String`",
         "errorCount": "Количество отображаемых сообщений об ошибках",
         "parse": "Функция, преобразующая текст поля ввода в дату, массив дат или период дат",
-        "parse-date": "Функция, преобразующая текст поля ввода в дату",
-        "parse-value": "Функция, преобразующая строковое значение в дату, массив дат или период дат"
+        "parseDate": "Функция, преобразующая текст поля ввода в дату",
+        "parseValue": "Функция, преобразующая строковое значение в дату, массив дат или период дат"
       },
       "events": {
         "updateView": "Событие запускается, когда компоненту необходимо изменить `view`",
