+++ packages/ui/src/components/va-date-input/hooks/model-value-parser.ts
@@ -2,19 +2,15 @@ import isNil from 'lodash/isNil'
 
 import { VaDatePickerModelValue } from '../../va-date-picker/types'
 import { defaultParseDateFunction } from './input-text-parser'
-import { VaDateInputRange, VaDateInputModelValue } from '../types'
+import { VaDateInputRange, VaDateInputModelValue, VaDateInputDate } from '../types'
 
 export const isRange = (date: VaDateInputModelValue): date is VaDateInputRange => {
   return typeof date === 'object' && ('start' in date || 'end' in date)
 }
 
-export const parseSingleDate = (date: string | number | Date) => {
-  if (typeof date === 'string') {
-    return defaultParseDateFunction(date)
-  }
-
-  return new Date(date)
-}
+export const parseSingleDate = (date: VaDateInputDate) => typeof date === 'string'
+  ? defaultParseDateFunction(date)
+  : new Date(date)
 
 export const parseModelValue = (date: VaDateInputModelValue): VaDatePickerModelValue => {
   if (Array.isArray(date)) {
