+++ packages/ui/src/components/va-date-picker/hooks/use-picker.ts
@@ -1,6 +1,6 @@
-import { isDatesMonthEqual, isDatesDayEqual, isDatesYearEqual } from './../utils/date-utils'
+import { isDatesMonthEqual, isDatesDayEqual, isDatesYearEqual, isDates, isRange, isSingleDate } from './../utils/date-utils'
 import { VaDatePickerMode, VaDatePickerModelValue, VaDatePickerType } from '../types'
-import { isDates, isRange, isSingleDate, useDatePickerModelValue } from './model-value-helper'
+import { useDatePickerModelValue } from './model-value-helper'
 import { computed, ComputedRef, ref, watch } from 'vue'
 
 const getDateEqualFunction = (type: VaDatePickerType) => {
