+++ packages/ui/src/components/va-date-picker/utils/date-utils.ts
@@ -1,12 +1,13 @@
+import { VaDatePickerModelValue, VaDatePickerRange } from '../types'
 
 export const isDatesDayEqual = (date1?: Date | null, date2?: Date | null): boolean => date1?.toDateString() === date2?.toDateString()
 
-export const isDatesMonthEqual = (date1?: Date | null, date2?: Date | null): boolean => {
-  return date1?.getFullYear() === date2?.getFullYear() && date1?.getMonth() === date2?.getMonth()
-}
-
 export const isDatesYearEqual = (date1?: Date | null, date2?: Date | null): boolean => {
   return date1?.getFullYear() === date2?.getFullYear()
+}
+
+export const isDatesMonthEqual = (date1?: Date | null, date2?: Date | null): boolean => {
+  return isDatesYearEqual(date1, date2) && date1?.getMonth() === date2?.getMonth()
 }
 
 export const createYearDate = (year: number) => {
@@ -22,3 +23,11 @@ export const isDatesArrayIncludeDay = (dates: Date[], date: Date): boolean => {
 export const isDatesArrayIncludeMonth = (dates: Date[], date: Date): boolean => {
   return !!dates.find((d) => isDatesMonthEqual(d, date))
 }
+
+export const isRange = (value: VaDatePickerModelValue): value is VaDatePickerRange => {
+  return typeof value === 'object' && ('start' in value || 'end' in value)
+}
+
+export const isSingleDate = (value: VaDatePickerModelValue): value is Date => value instanceof Date
+
+export const isDates = (value: VaDatePickerModelValue): value is Date[] => Array.isArray(value)
